@@ -8,50 +8,28 @@
     "build": "NODE_ENV=production npx webpack --mode production --progress",
     "build:dev": "npx webpack --mode development",
     "prettify": "yarn prettier \"**/*.*(js|ts|tsx)\" --ignore-path=.prettierignore --write",
-    "lint": "eslint src --ext .ts,.tsx && yarn prettify",
     "test": "jest --coverage"
-  },
-  "lint-staged": {
-    "*.{js,ts,tsx}": [
-      "eslint --fix",
-      "prettier --write"
-    ]
   },
   "dependencies": {
     "@usedapp/core": "^1.0.2",
     "mobx": "^6.6.0",
     "mobx-react": "^7.5.0",
-    "react": "^17.0.1",
-    "react-color": "^2.19.3",
-    "react-dom": "^17.0.1",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-router-dom": "^6.3.0",
     "react-zoom-pan-pinch": "^2.1.3",
     "sass": "^1.51.0"
   },
   "devDependencies": {
-    "@testing-library/react": "^11.2.2",
-    "@types/react": "^17.0.0",
-    "@types/react-color": "^3.0.6",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router-dom": "^5.1.6",
+    "@types/react": "^18.0.11",
+    "@types/react-dom": "^18.0.5",
+    "@types/react-router-dom": "^5.3.3",
     "@types/webpack": "^4.41.25",
-    "@typescript-eslint/eslint-plugin": "^4.11.1",
-    "@typescript-eslint/parser": "^4.10.0",
     "clean-webpack-plugin": "^3.0.0",
     "css-loader": "^6.7.1",
-    "eslint": "^7.15.0",
-    "eslint-config-prettier": "^7.1.0",
-    "eslint-import-resolver-typescript": "^2.3.0",
-    "eslint-import-resolver-webpack": "^0.13.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest": "^24.1.3",
-    "eslint-plugin-react": "^7.21.5",
-    "eslint-plugin-react-hooks": "^4.2.0",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "5.0.0-alpha.3",
     "jest": "^26.6.3",
-    "lint-staged": "^10.5.3",
     "node-sass": "^7.0.1",
     "prettier": "^2.2.1",
     "sass-loader": "^12.6.0",
