@@ -1,13 +1,11 @@
-import React from "react";
-import "../styles/pixel.scss";
+import React from 'react';
+import '../styles/pixel.scss';
 
 type PixelProps = {
-    color: string
-}
+  color: string;
+};
 
 export const Pixel = (props: PixelProps) => {
-  const { color } = props
-  return (
-        <div></div>
-  );
-}
+  const { color } = props;
+  return <div></div>;
+};
