@@ -1,10 +1,10 @@
-import { Mainnet } from "@usedapp/core"
+import { Mainnet } from '@usedapp/core';
 
 const config = {
-    chainId: Mainnet.chainId,
-    rpcUrl: "",
-    imageDimensions: 30,
-    apiUrl: ""
-}
+  chainId: Mainnet.chainId,
+  rpcUrl: 'http://127.0.0.1:8545',
+  imageDimensions: 30,
+  apiUrl: 'http://localhost:8080',
+};
 
-export default config
+export default config;
