@@ -1,12 +1,12 @@
 export type Pixel = {
-    x: number 
-    y: number 
-    color: string 
-    editor: string
-    updatedAt: string
-}
+  x: number;
+  y: number;
+  color: string;
+  editor: string;
+  updatedAt: string;
+};
 
 export type SimplePixel = {
-    x: number 
-    y: number
-}
+  x: number;
+  y: number;
+};
